python-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pyodbc==4.0.39
python-dotenv==1.0.0
Pillow==10.0.0
